+++ cmd/app.go
@@ -2,6 +2,7 @@
 package cmd
 
 import (
+	"bytes"
 	"context"
 	"encoding/json"
 	"io"
@@ -79,6 +80,16 @@ func (sub *substation) SetConfig(r io.Reader) error {
 	}
 
 	return nil
+}
+
+// Config retreives the configuration of the app.
+func (sub *substation) Config() (io.Reader, error) {
+	var buf bytes.Buffer
+	if err := json.NewEncoder(&buf).Encode(sub.config); err != nil {
+		return nil, err
+	}
+
+	return &buf, nil
 }
 
 // Concurrency returns the concurrency setting of the app.